+++ core/src/main/java/org/musetest/core/step/ContainsStep.java
@@ -1,9 +1,11 @@
 package org.musetest.core.step;
 
+import org.musetest.core.*;
+
 /**
  * @author Christopher L Merrill (see LICENSE.txt for license details)
  */
-public interface ContainsStep
+public interface ContainsStep extends MuseResource
     {
     StepConfiguration getStep();
     }