+++ core/src/main/java/org/musetest/core/suite/IdListTestSuite.java
@@ -21,7 +21,7 @@ public class IdListTestSuite implements MuseTestSuite
         List<TestConfiguration> tests = new ArrayList<>(_test_ids.size());
         for (String id : _test_ids)
             {
-            MuseTest test = project.findResource(id, MuseTest.class);
+            MuseTest test = project.getResource(project.findResource(id, MuseTest.class));
             if (test == null)
                 {
                 LOG.error("Test with id {} was not found in the project", id);